Enforce most of the server actions auth
Also adds a toast that shows the result of a request

+++ src/actions/banners.ts
@@ -6,6 +6,15 @@ import { redirect } from 'next/navigation';
 import { MediaType } from '@/services/fileService';
 
 export async function addBanner(divisionGroupId: number, form: FormData) {
+  if (
+    !(
+      (await SessionService.isAdmin()) ||
+      (await SessionService.canEditGroupByInternalId(divisionGroupId))
+    )
+  ) {
+    throw new Error('Unauthorized');
+  }
+
   const file: File | null = form.get('file') as unknown as File;
 
   if (!file) {
@@ -20,6 +29,20 @@ export async function addBanner(divisionGroupId: number, form: FormData) {
 }
 
 export async function deleteBanner(bannerId: number) {
+  const divisionGroupId = await DivisionGroupService.getBannerOwner(bannerId);
+  if (divisionGroupId === undefined) {
+    throw new Error('Banner owner not found');
+  }
+
+  if (
+    !(
+      (await SessionService.isAdmin()) ||
+      (await SessionService.canEditGroupByInternalId(divisionGroupId))
+    )
+  ) {
+    throw new Error('Unauthorized');
+  }
+
   await DivisionGroupService.deleteBanner(bannerId);
   redirect('./banners');
 }
